# docs/swagger.yaml
basePath: /api/public/accounting/v1
definitions:
...
  dto.CreateBookRequest:
    properties:
      base_currency_code:
        example: INR
        type: string
      book_type:
        enum:
        - PRIMARY
        - STATUTORY
        - MANAGEMENT
        example: PRIMARY
        type: string
      code:
        example: PRIMARY
        maxLength: 30
        type: string
      company_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      name:
        example: Primary Book
        maxLength: 120
        type: string
      reporting_currency_code:
        example: USD
        type: string
    required:
    - base_currency_code
    - book_type
    - code
    - company_id
    - name
    type: object
  dto.CreateFiscalYearRequest:
    properties:
      book_id:
        example: 550e8400-e29b-41d4-a716-446655440010
        type: string
      code:
        example: FY2026
        maxLength: 30
        type: string
      company_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      end_date:
        example: "2027-03-31T23:59:59Z"
        type: string
      name:
        example: FY 2026-27
        maxLength: 200
        type: string
      start_date:
        example: "2026-04-01T00:00:00Z"
        type: string
      status:
        enum:
        - DRAFT
        - OPEN
        - CLOSED
        - ARCHIVED
        example: DRAFT
        type: string
    required:
    - book_id
    - code
    - company_id
    - end_date
    - name
    - start_date
    type: object
...
  dto.UpdateBookRequest:
    properties:
      name:
        example: Primary Book Updated
        maxLength: 120
        type: string
      reporting_currency_code:
        example: USD
        type: string
      status:
        enum:
        - ACTIVE
        - INACTIVE
        example: ACTIVE
        type: string
    type: object
  dto.UpdateFiscalYearRequest:
    properties:
      end_date:
        example: "2027-03-31T23:59:59Z"
        type: string
      name:
        example: FY 2026-27 (revised)
        maxLength: 200
        type: string
      status:
        enum:
        - DRAFT
        - OPEN
        - CLOSED
        - ARCHIVED
        example: OPEN
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      error:
        items:
          type: string
        type: array
      message:
        type: string
      status:
        type: boolean
      status_code:
        type: integer
    type: object
host: localhost:8086
info:
  contact:
    email: support@queuebuster.co
    name: API Support
    url: http://www.queuebuster.co/support
  description: Enterprise accounting module for QueueBuster ERP
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: QB Accounting Service API
  version: 1.0.0
paths:
...
  /account-groups/book/{book_id}:
    get:
      description: List account groups for a book
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List account groups by book
      tags:
      - account-groups
...
  /accounting-periods/book/{book_id}:
    get:
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List accounting periods by book
      tags:
      - accounting-periods
  /accounting-periods/fiscal-year/{fiscal_year_id}:
    get:
      parameters:
      - description: Fiscal year ID
        in: path
        name: fiscal_year_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List accounting periods by fiscal year
      tags:
      - accounting-periods
...
  /accounts/book/{book_id}:
    get:
      description: List accounts for an accounting book
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List accounts by book
      tags:
      - accounts
...
  /books:
    post:
      consumes:
      - application/json
      description: Create a new accounting book for a company
      parameters:
      - description: Book data
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/dto.CreateBookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a book
      tags:
      - books
  /books/{id}:
    delete:
      description: Delete an accounting book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Delete a book
      tags:
      - books
    get:
      description: Get an accounting book by ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a book
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Update an accounting book
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Book fields
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateBookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update a book
      tags:
      - books
  /books/company/{company_id}:
    get:
      description: List all accounting books for a company
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List books by company
      tags:
      - books
  /fiscal-years:
    post:
      consumes:
      - application/json
      description: Create a new fiscal year for a book
      parameters:
      - description: Fiscal year data
        in: body
        name: fiscal_year
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFiscalYearRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a fiscal year
      tags:
      - fiscal-years
  /fiscal-years/{id}:
    delete:
      description: Delete a fiscal year
      parameters:
      - description: Fiscal year ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Delete a fiscal year
      tags:
      - fiscal-years
    get:
      description: Get a fiscal year by ID
      parameters:
      - description: Fiscal year ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a fiscal year
      tags:
      - fiscal-years
    put:
      consumes:
      - application/json
      description: Update a fiscal year
      parameters:
      - description: Fiscal year ID
        in: path
        name: id
        required: true
        type: string
      - description: Fiscal year fields
        in: body
        name: fiscal_year
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateFiscalYearRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update a fiscal year
      tags:
      - fiscal-years
  /fiscal-years/book/{book_id}:
    get:
      description: List fiscal years for an accounting book
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List fiscal years by book
      tags:
      - fiscal-years
...
  /journal-batches/book/{book_id}:
    get:
      description: List journal batches for a book
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List journal batches by book
      tags:
      - journal-batches
...
  /journals/book/{book_id}:
    get:
      description: List journals for an accounting book
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List journals by book
      tags:
      - journals
...
  /open-items/book/{book_id}:
    get:
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List open items by book
      tags:
      - open-items
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"